spec: use macrocache to store total test count
nimskull does not allow compileTime variable to be reached from
potentially runtime code.

While this issue could be solved by adding compileTime to totalTests,
doing so breaks the variable due to
https://github.com/nim-works/nimskull/issues/1022.

Instead, store the counter using macrocache, which avoids this issue
rather nicely (and is the recommended way in Nim to store these values).